fix: lowercase folders
