+++ src/policies/DefaultPolicy.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Lmr\AutoTranslator\Policies;
+namespace Lmr\AutoTranslator\policies;
 
 use craft\elements\Entry;
 use craft\helpers\ElementHelper;
-use Lmr\AutoTranslator\Contracts\PolicyInterface;
+use Lmr\AutoTranslator\contracts\PolicyInterface;
 use Lmr\AutoTranslator\Plugin;
 
 class DefaultPolicy implements PolicyInterface
